Move player designated zones into player class/their own files.
What was this doing in map.h?
